Remove 'word' element from HtmlBuilder
Styling using the word element causes added whitespace whenever a
ChordLyricsPair starts with a space and when the lyrics is shorter than
the chord. For instance

```
            A#m      F#
It's a small    crime,
        D#m7            F#
All the  angels and the Saints
```

instead of

```
            A#m   F#
It's a small crime,
        D#m7            F#
All the angels and the Saints
```

As a matter of personal preference, I have changed the code to produce
the latter case.